+++ crates/eunomia/src/impls/mod.rs
@@ -1,3 +1,4 @@
 mod field;
 mod primitives;
+mod unit;
 mod wrappers;